+++ resources/lang/en/fileuploadcare-component.php
@@ -1,4 +1,0 @@
-<?php
-
-// translations for Vormkracht10/FileUploadcare
-return [];