+++ resources/lang/en/uploadcare-field.php
@@ -1,0 +1,4 @@
+<?php
+
+// translations for Vormkracht10/Uploadcare
+return [];