+++ src/Facades/Uploadcare.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Vormkracht10\Uploadcare\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+/**
+ * @see \Vormkracht10\Uploadcare\Uploadcare
+ */
+class Uploadcare extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return \Vormkracht10\Uploadcare\Uploadcare::class;
+    }
+}