+++ src/FileUploadcare.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Vormkracht10\FileUploadcare;
-
-class FileUploadcare {}