+++ src/Uploadcare.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Vormkracht10\Uploadcare;
+
+class Uploadcare {}