+++ tests/TestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Vormkracht10\FileUploadcare\Tests;
+namespace Vormkracht10\Uploadcare\Tests;
 
 use BladeUI\Heroicons\BladeHeroiconsServiceProvider;
 use BladeUI\Icons\BladeIconsServiceProvider;
@@ -16,7 +16,7 @@ use Illuminate\Database\Eloquent\Factories\Factory;
 use Livewire\LivewireServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 use RyanChandler\BladeCaptureDirective\BladeCaptureDirectiveServiceProvider;
-use Vormkracht10\FileUploadcare\FileUploadcareServiceProvider;
+use Vormkracht10\Uploadcare\UploadcareServiceProvider;
 
 class TestCase extends Orchestra
 {
@@ -25,7 +25,7 @@ class TestCase extends Orchestra
         parent::setUp();
 
         Factory::guessFactoryNamesUsing(
-            fn (string $modelName) => 'Vormkracht10\\FileUploadcare\\Database\\Factories\\' . class_basename($modelName) . 'Factory'
+            fn(string $modelName) => 'Vormkracht10\\Uploadcare\\Database\\Factories\\' . class_basename($modelName) . 'Factory'
         );
     }
 
@@ -44,7 +44,7 @@ class TestCase extends Orchestra
             SupportServiceProvider::class,
             TablesServiceProvider::class,
             WidgetsServiceProvider::class,
-            FileUploadcareServiceProvider::class,
+            UploadcareServiceProvider::class,
         ];
     }
 
@@ -53,7 +53,7 @@ class TestCase extends Orchestra
         config()->set('database.default', 'testing');
 
         /*
-        $migration = include __DIR__.'/../database/migrations/create_filament-fileuploadcare-component_table.php.stub';
+        $migration = include __DIR__.'/../database/migrations/create_filament-uploadcare-field_table.php.stub';
         $migration->up();
         */
     }