+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
-All notable changes to `filament-fileuploadcare-component` will be documented in this file.
+All notable changes to `filament-uploadcare-field` will be documented in this file.
 
 ## 1.0.0 - 202X-XX-XX
 
-- initial release
+-   initial release